refactor: 자동 로또 생성기 패키지 변경
AutoLottoGenerator.java
InputHandler
InputHandlerErrorMessage.java
LottoController.java
